upd: reorg/condense upd: clean up error messages
